apps/spot/requirements.txt
python-dotenv
openai
neurosity
opencv-python
requests
setuptools
elevenlabs
tenacity